mcp-servers/requirements.txt
django>=5.1,<5.2
django-cors-headers>=4.4.0
dedalus-labs
httpx>=0.27.0
python-dotenv>=1.0.0
